\name{MRDA2018-package}
\alias{MRDA2018-package}
\alias{MRDA2018}
\docType{package}
\title{\packageTitle{MRDA2018}}
\description{\packageDescription{MRDA2018}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{MRDA2018}
  \packageIndices{MRDA2018}
 Simply load the package to generate the data for the MRDA2018 examples with dplyr. 
}
\author{
  \packageAuthor{MRDA2018}
  
  Maintainer: \packageMaintainer{MRDA2018}
}

